+++ tests/testbed/019-rbi-is-a-declaration/app.rb
@@ -1,0 +1,12 @@
+class Job
+  def run
+    w = Widget.new
+    w.size
+    w.weigh
+  end
+end
+
+class Widget
+  def weigh
+  end
+end